Further refactoring and review comment fixes in account create/update API
